+++ modules/engine/test/index.ts
@@ -42,3 +42,4 @@ import './compute/texture-transform.spec';
 import './compute/computation.spec';
 import './passes/shader-pass-renderer.spec';
 import './models/background-texture-model.spec';
+import './models/light-models.spec';