+++ website/src/examples.tsx
@@ -133,11 +133,7 @@ export const FP64Example: React.FC = () => {
   }
 
   return deviceType && presentationDevice ? (
-    <ReactExample
-      component={FP64App}
-      componentProps={{presentationDevice}}
-      showStats={false}
-    />
+    <ReactExample component={FP64App} componentProps={{presentationDevice}} showStats={false} />
   ) : (
     <ExamplePage>
       <div>Initializing device...</div>
@@ -235,7 +231,6 @@ export const HelloTriangleExample: React.FC = props => (
     directory="tutorials"
     template={HelloTriangleApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -247,7 +242,6 @@ export const HelloTriangleGeometryExample: React.FC = props => (
     directory="tutorials"
     template={HelloTriangleGeometryApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -259,7 +253,6 @@ export const HelloCubeExample: React.FC = props => (
     directory="tutorials"
     template={HelloCubeApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -271,7 +264,6 @@ export const InstancedCubesExample: React.FC = props => (
     directory="tutorials"
     template={InstancedCubesApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -283,7 +275,6 @@ export const TwoCubesExample: React.FC = props => (
     directory="tutorials"
     template={TwoCubesApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -295,7 +286,6 @@ export const LightingExample: React.FC = props => (
     directory="tutorials"
     template={LightingApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -307,7 +297,6 @@ export const HelloGLTFExample: React.FC = props => (
     directory="tutorials"
     template={HelloGLTFApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -319,7 +308,6 @@ export const HelloInstancingExample: React.FC = props => (
     directory="tutorials"
     template={HelloInstancingApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -331,7 +319,6 @@ export const ShaderHooksExample: React.FC = props => (
     directory="tutorials"
     template={ShaderHooksApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -343,7 +330,6 @@ export const ShaderModulesExample: React.FC = props => (
     directory="tutorials"
     template={ShaderModulesApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -355,7 +341,6 @@ export const TransformFeedbackExample: React.FC = props => (
     directory="tutorials"
     template={TransformFeedbackApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />
@@ -367,7 +352,6 @@ export const TransformExample: React.FC = props => (
     directory="tutorials"
     template={TransformApp}
     config={exampleConfig}
-    showHeader={false}
     showStats={false}
     {...props}
   />